dev-requirements.txt
pre-commit==2.4.0
pylint==2.4.4
prospector==1.2.0

pytest==5.3.5
pytest-selenium==1.17.0

